+++ apps/pioneer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @internal/playground
 
+## 0.4.12
+
+### Patch Changes
+
+- bump
+
+## 0.4.11
+
+### Patch Changes
+
+- bump
+
+## 0.4.10
+
+### Patch Changes
+
+- bump
+
 ## 0.4.9
 
 ### Patch Changes