+++ e2e/wallets/intergration-wallet-keystore/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @coinmasters/integration-wallet-keepkey
 
+## 1.4.12
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.4.12
+  - @coinmasters/wallet-keystore@8.0.12
+  - @coinmasters/wallet-keepkey@10.0.13
+  - @coinmasters/types@4.4.12
+  - @coinmasters/core@8.0.12
+
+## 1.4.11
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.4.11
+  - @coinmasters/wallet-keystore@8.0.11
+  - @coinmasters/wallet-keepkey@10.0.12
+  - @coinmasters/types@4.4.11
+  - @coinmasters/core@8.0.11
+
+## 1.4.10
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.4.10
+  - @coinmasters/wallet-keystore@8.0.10
+  - @coinmasters/wallet-keepkey@10.0.11
+  - @coinmasters/types@4.4.10
+  - @coinmasters/core@8.0.10
+
 ## 1.4.9
 
 ### Patch Changes