+++ packages/swapkit/entities/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @thorswap-lib/swapkit-entities
 
+## 3.4.12
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.4.12
+
+## 3.4.11
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.4.11
+
+## 3.4.10
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.4.10
+
 ## 3.4.9
 
 ### Patch Changes