+++ packages/wallets/okx/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @coinmasters/wallet-okx
 
+## 8.0.12
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@8.0.12
+  - @coinmasters/helpers@8.0.12
+  - @coinmasters/toolbox-utxo@8.0.12
+  - @coinmasters/types@4.4.12
+  - @coinmasters/toolbox-evm@8.0.12
+
+## 8.0.11
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@8.0.11
+  - @coinmasters/helpers@8.0.11
+  - @coinmasters/toolbox-utxo@8.0.11
+  - @coinmasters/types@4.4.11
+  - @coinmasters/toolbox-evm@8.0.11
+
+## 8.0.10
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@8.0.10
+  - @coinmasters/helpers@8.0.10
+  - @coinmasters/toolbox-utxo@8.0.10
+  - @coinmasters/types@4.4.10
+  - @coinmasters/toolbox-evm@8.0.10
+
 ## 8.0.9
 
 ### Patch Changes